--- src/components/QuestionTopBar.jsx
@@ -1,22 +1,0 @@
-import React from "react";
-
-export default function QuestionTopBar(){
-  return (
-    <div className="sticky top-0 z-20 bg-white">
-      <div className="mx-auto max-w-[760px] flex items-center justify-between h-12 px-3">
-        <button aria-label="뒤로가기" className="p-2 bg-white ">
-          <svg xmlns="http://www.w3.org/2000/svg" width="10" height="16" viewBox="0 0 10 16" fill="none">
-            <path fillRule="evenodd" clipRule="evenodd" d="M1.15017 8.8295L7.75 15.4293L9.39967 13.7797L3.62467 8.00467L9.39967 2.22967L7.75 0.58L1.15017 7.17983C0.931451 7.39862 0.808583 7.69531 0.808583 8.00467C0.808583 8.31402 0.931451 8.61072 1.15017 8.8295Z" fill="#191D1F"/>
-          </svg>
-        </button>
-        <div className="flex items-center gap-1">
-          <button aria-label="메뉴" className="p-2 rounded-full hover:bg-neutral-100">
-            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-              <path d="M3 7H21C21.2652 7 21.5196 6.89464 21.7071 6.70711C21.8946 6.51957 22 6.26522 22 6C22 5.73478 21.8946 5.48043 21.7071 5.29289C21.5196 5.10536 21.2652 5 21 5H3C2.73478 5 2.48043 5.10536 2.29289 5.29289C2.10536 5.48043 2 5.73478 2 6C2 6.26522 2.10536 6.51957 2.29289 6.70711C2.48043 6.89464 2.73478 7 3 7ZM21 11H3C2.73478 11 2.48043 11.1054 2.29289 11.2929C2.10536 11.4804 2 11.7348 2 12C2 12.2652 2.10536 12.5196 2.29289 12.7071C2.48043 12.8946 2.73478 13 3 13H21C21.2652 13 21.5196 12.8946 21.7071 12.7071C21.8946 12.5196 22 12.2652 22 12C22 11.7348 21.8946 11.4804 21.7071 11.2929C21.5196 11.1054 21.2652 11 21 11ZM21 17H3C2.73478 17 2.48043 17.1054 2.29289 17.2929C2.10536 17.4804 2 17.7348 2 18C2 18.2652 2.10536 18.5196 2.29289 18.7071C2.48043 18.8946 2.73478 19 3 19H21C21.2652 19 21.5196 18.8946 21.7071 18.7071C21.8946 18.5196 22 18.2652 22 18C22 17.7348 21.8946 17.4804 21.7071 17.2929C21.5196 17.1054 21.2652 17 21 17Z" fill="#191D1F"/>
-            </svg>
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
